// blog-frontend/src/types/api.ts
export interface ApiResponse<T = any> {
  code: number
  msg?: string
  message?: string
  data: T
}